Split AlleleCountMatrix into its own hpp/cc files.
