+++ Sequence/Makefile.in
@@ -391,6 +391,7 @@ pkginclude_HEADERS = AlignStream.hpp\
 	SeqAlphabets.hpp \
 	VariantMatrix.hpp \
 	VariantMatrixViews.hpp \
+	AlleleCountMatrix.hpp \
 	summstats.hpp \
 	StateCounts.hpp
