+++ Sequence/VariantMatrix.hpp
@@ -120,23 +120,8 @@ namespace Sequence
         /// std::out_of_range is thrown if indexes are invalid.
         const std::int8_t& at(const std::size_t site,
                               const std::size_t haplotype) const;
-
-        //AlleleCountMatrix count_alleles() const;
     };
 
-    class AlleleCountMatrix
-    /// \brief Matrix representation of allele counts in a VariantMatrix
-    /// To be constructed
-    {
-      private:
-        static std::vector<std::int32_t> init_counts(const VariantMatrix& m);
-
-      public:
-        const std::vector<std::int32_t> counts;
-        const std::size_t row_size;
-        const std::size_t nsam;
-        explicit AlleleCountMatrix(const VariantMatrix& m);
-    };
 } // namespace Sequence
 
 #endif
