+++ Sequence/summstats/thetapi.hpp
@@ -4,7 +4,7 @@
 #ifndef SEQUENCE_SUMMSTATS_THETAPI_HPP__
 #define SEQUENCE_SUMMSTATS_THETAPI_HPP__
 
-#include <Sequence/VariantMatrix.hpp>
+#include <Sequence/AlleleCountMatrix.hpp>
 
 namespace Sequence
 {
