+++ Sequence/summstats/thetaw.hpp
@@ -3,7 +3,7 @@
 #ifndef SEQUENCE_SUMMSTATS_THETAW_HPP__
 #define SEQUENCE_SUMMSTATS_THETAW_HPP__
 
-#include <Sequence/VariantMatrix.hpp>
+#include <Sequence/AlleleCountMatrix.hpp>
 
 namespace Sequence
 {
